+++ tests/snapshot_tests/test_snapshots.py
@@ -176,4 +176,8 @@ def test_key_display(snap_compare):
 
 def test_demo(snap_compare):
     """Test the demo app (python -m textual)"""
-    assert snap_compare(Path("../../src/textual/demo.py"))
+    assert snap_compare(
+        Path("../../src/textual/demo.py"),
+        press=["down", "down", "down", "_"],
+        terminal_size=(100, 30),
+    )